feat: commonService, Dao 생성, 코드 이동
