+++ dao/user.js
@@ -209,19 +209,6 @@ const checkToken = async (refreshToken) => {
   }
 };
 
-const getLocation = async (user_id) => {
-  const conn = await pool.getConnection();
-  try {
-    const checkSql = 'SELECT latitude, longitude, sigungu FROM user WHERE id = ?';
-    const [checkRows] = await conn.query(checkSql, user_id);
-    return checkRows;
-  } catch (err) {
-    throw customError(500, err.sqlMessage);
-  } finally {
-    await conn.release();
-  }
-};
-
 module.exports = {
   signup,
   login,
@@ -234,5 +221,4 @@ module.exports = {
   verifiedCheck,
   emailVerificationProcess,
   checkToken,
-  getLocation,
 };
